[refactor] 일급컬렉션 적용
